# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: VouckBack
  description: |-
    Microserviço para geração de voucher como descontos em valor ou percentual paras clientes em seu ponto de venda.

    _Se você que saber mais sobre a TIBBI, então  click [aqui](http://tibbi.com.br)._

  contact:
    email: mario7lagoas@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/voucher-back/v1
...
paths:
...
  /perfil:
    get:
      tags:
        - Perfil
      summary: Buscando Lista de Perfil
      description: Buscando Lista de Perfil na Plataforma
      operationId: buscandoListaPerfil
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerfilApiResponse'
    post:
      tags:
        - Perfil
      summary: Criando Um Perfil
      description: Criando Um Perfil na Plataforma
      operationId: criandoPerfil
      requestBody:
        description: Criando novo Perfil
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerfilApiRequest'
        required: true
      responses:
        '201':
          description: created successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilApiResponse'
        '400':
          description: bad request operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 400 BAD_REQUEST
                  erros:
                    type: array
                    items:
                      properties:
                        mensagem:
                          type: string
                          example: "não deve ser nulo"
                        campo:
                          type: string
                          example: "nome"
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Perfil já cadastrado.
                        mensagem:
                          type: string
                          example: Já existe um Perfil com este nome.
...
  /perfil/{guid}:
    get:
      tags:
        - Perfil
      summary: Buscando Perfil Pelo Perfil GUID
      description: Buscando Perfil Pelo Perfil GUID na Plataforma
      operationId: buscandoPerfilPeloGUID
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: Perfil não encontrado.
    put:
      tags:
        - Perfil
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          example: aaaa-bbbb-cccc-dddd-eeeee
          required: true
          schema:
            type: string
      summary: Alterando um Perfil
      description: Alterando um Perfil na Plataforma
      operationId: alterandoPerfil
      requestBody:
        description: Alterando Perfil Existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerfilUpdateApiRequest'
      responses:
        '202':
          description: accepted successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: Perfil não encontrado.
    delete:
      tags:
        - Perfil
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      summary: Apagando um Perfil
      description: Apagando um Perfil na Plataforma
      operationId: apagandoPerfil
      responses:
        '204':
          description: No Content successful operation
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: Perfil não encontrado.
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não permitido Excluir.
                        mensagem:
                          type: string
                          example: Não permitido Excluir. Perfil associado a algum Usuario.
  /perfil/{nome}/nome:
    get:
      tags:
        - Perfil
      summary: Buscando Perfil Pelo Perfil Nome
      description: Buscando Perfil Pelo Perfil Nome na Plataforma
      operationId: buscandoPerfilPeloNome
      parameters:
        - name: nome
          in: path
          description: Nome do Perfil
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: Perfil não encontrado.
...
components:
  schemas:
...
    PerfilApiRequest:
      required:
        - nome
        - roles
      type: object
      properties:
        nome:
          type: string
          description: "Nome do Perfil"
          example: Moderator
        roles:
          type: array
          description: "Permissões do Perfil"
          items:
            $ref: '#/components/schemas/RoleApiResponse'
    PerfilUpdateApiRequest:
      type: object
      properties:
        nome:
          type: string
          description: "Nome do Perfil"
          example: Moderator
        roles:
          type: array
          description: "Permissões do Perfil"
          items:
            $ref: '#/components/schemas/RoleApiResponse'
...
    PerfilApiResponse:
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
        nome:
          type: string
          example: "Moderator"
          description: "Nome do Perfil"
        roles:
          type: array
          description: "Permissões do Perfil"
          items:
            $ref: '#/components/schemas/RoleApiResponse'
    RoleApiResponse:
      type: object
      properties:
        nome:
          type: string
          example: "MENU_LOJA"
          description: "Permissão"
...